+++ app/smartbrain_3000/web/_app/immutable/entry/start.dRA44zV0.js
@@ -1,0 +1,1 @@
+import{a as e,n as t}from"../chunks/BtL57GHQ.js";export{e as load_css,t as start};